SNOW-2043816: Customization tests running
